Rename api/propagators to api/propagation
This package will not contain any propagators in the long run, just
the interface definitions.
