+++ api/propagation/doc.go
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package propagators contains interface definition for BinaryFormat and
+// Package propagation contains interface definition for BinaryFormat and
 // TextFormat propagators.
-package propagators // import "go.opentelemetry.io/otel/api/propagators"
+package propagation // import "go.opentelemetry.io/otel/api/propagation"
