+++ api/propagation/text_noop_propagator.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package propagators
+package propagation
 
 import (
 	"context"
