+++ plugin/othttp/handler.go
@@ -20,7 +20,7 @@ import (
 
 	"go.opentelemetry.io/otel/api/core"
 	"go.opentelemetry.io/otel/api/global"
-	"go.opentelemetry.io/otel/api/propagators"
+	"go.opentelemetry.io/otel/api/propagation"
 	"go.opentelemetry.io/otel/api/trace"
 )
 
@@ -50,7 +50,7 @@ type Handler struct {
 	handler   http.Handler
 
 	tracer           trace.Tracer
-	prop             propagators.TextFormat
+	prop             propagation.TextFormat
 	spanStartOptions []trace.StartOption
 	public           bool
 	readEvent        bool
@@ -80,7 +80,7 @@ func WithPublicEndpoint() Option {
 // WithPropagator configures the Handler with a specific propagator. If this
 // option isn't specificed then
 // go.opentelemetry.io/otel/api/trace.TraceContext is used.
-func WithPropagator(p propagators.TextFormat) Option {
+func WithPropagator(p propagation.TextFormat) Option {
 	return func(h *Handler) {
 		h.prop = p
 	}
